:recycle: Move socket events into separate socket manager file
